# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Course Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/checkout:
    post:
      operationId: createCheckout
      tags: [payments]
      summary: Create payment checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutBody"
      responses:
        "200":
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify payment and enroll user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: Payment verified, enrollment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
...
components:
  schemas:
...
    CreateCheckoutBody:
      type: object
      properties:
        courseId:
          type: integer
        couponCode:
          type: string
          nullable: true
        gateway:
          type: string
          enum: [stripe, razorpay]
      required: [courseId, gateway]
    CheckoutResponse:
      type: object
      properties:
        sessionId:
          type: string
        amount:
          type: number
        currency:
          type: string
        gateway:
          type: string
        redirectUrl:
          type: string
          nullable: true
        razorpayOrderId:
          type: string
          nullable: true
        razorpayKey:
          type: string
          nullable: true
      required: [sessionId, amount, currency, gateway]
    VerifyPaymentBody:
      type: object
      properties:
        sessionId:
          type: string
        paymentId:
          type: string
          nullable: true
        signature:
          type: string
          nullable: true
      required: [sessionId]
    PaymentVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
        enrollmentId:
          type: integer
          nullable: true
        message:
          type: string
      required: [success, message]
...